requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.0.0
tqdm>=4.66.0
Pillow>=10.0.0
nltk>=3.8.0
rouge>=1.0.1
scikit-learn>=1.3.0
git+https://github.com/openai/CLIP.git  # OpenAI CLIP vision encoder
